--- Assets/LeapMotionModules/FramerateGraph/README.txt.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3e3f2ea9d6f19694aa416c7f9eceefd5
-timeCreated: 1464306492
-licenseType: Pro
-TextScriptImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
