--- Assets/LeapMotionModules/FramerateGraph/Scripts/FramerateGraph.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 033555e4ae92b2b49a751156a005b649
-timeCreated: 1464302446
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
