+++ Assets/LeapMotionModules/RealtimeGraph/Material.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: d168ae295a9ee2b4d9c1da88e081b62e
-timeCreated: 1465947927
+guid: 0c9ee2af022180648bca3ca4f1259d60
+folderAsset: yes
+timeCreated: 1466116759
 licenseType: Pro
 DefaultImporter:
   userData:
