+++ Assets/LeapMotionModules/RealtimeGraph/Prefabs/RealtimeGraph.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: 83e697c78e750d241958b14ec457345f
-timeCreated: 1446583873
+timeCreated: 1466116470
 licenseType: Pro
 NativeFormatImporter:
   userData:
