+++ Assets/LeapMotionModules/RealtimeGraph/Textures.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 029485c5caa5d824888b489934e23f31
+folderAsset: yes
+timeCreated: 1466116736
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
